Complete Structural Re-Write of jounal_data.py...again.
A common theme among every major project is if the program handles files, I will, at consistent intervals re-work the file handling strucutre.  I swear 'this is the last time', but it never is.  But hey, LOOK AT THAT SWEET FLOW!

OK, it's not that sweet, but the previous file structure was really hard to follow, and most of the class methods relied on one of many boolean arguments that determined what that function should do.  Most classes were cleaned up through the implementation of the FileIO class, which serves as a mixin of sorts for all file read/write operations.  If a method needs to open a file, or write a file, it uses a FileIO method.  This cut down on spaghetti code issues.

Methods not replaced by FileIO were swapped around in some cases. JsonData got split up, and a new class, DataValidation, now exclusivley handles all file/directory path checks.  In other news, my excess use of @classmethod is going to bother someone.  I use it as an organizational technique, and creating instances of file handling methods is shenanigans.  Probably.

The only functional change comes from the use of the /end switch to end journal input.  It can now be used on short entires (where a 'new line' hasn't been entered yet, so there's something